initial split of ofx reader away from investments
